+++ backend/internal/data/ent/schema/entity.go
@@ -48,7 +48,7 @@ func (Entity) Fields() []ent.Field {
 			Default(false),
 		field.Bool("archived").
 			Default(false),
-		field.Int("asset_id").
+		field.Int64("asset_id").
 			Default(0),
 		field.Bool("sync_child_entity_locations").
 			Default(false),
